Details refactored 3

+++ app/src/main/java/com/example/myfilms/presentation/fragments/favorite/FavoritesFragment.kt
@@ -5,6 +5,7 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.activity.OnBackPressedCallback
+import androidx.appcompat.app.AlertDialog
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProvider
 import androidx.navigation.fragment.findNavController
@@ -85,11 +86,21 @@ class FavoritesFragment : Fragment() {
     private fun onBackPressed() {
         val callback = object : OnBackPressedCallback(true) {
             override fun handleOnBackPressed() {
-                try {
-                    viewModel.deleteSession()
-                    findNavController().popBackStack()
-                } catch (e: Exception) {
-                    findNavController().popBackStack()
+                requireContext().let {
+                    AlertDialog
+                        .Builder(it)
+                        .setMessage("Выйти?")
+                        .setPositiveButton("Да") { dialogInterface, i ->
+                            try {
+                                viewModel.deleteSession()
+                                findNavController().popBackStack()
+                            } catch (e: Exception) {
+                                findNavController().popBackStack()
+                            }
+                        }
+                        .setNegativeButton("Нет") { dialogInterface, i -> }
+                        .create()
+                        .show()
                 }
             }
         }
